+++ swan/dataset/geometry.py
@@ -2,14 +2,13 @@
 import json
 import multiprocessing
 from functools import partial
-from pathlib import Path
-from typing import Collection, List, Tuple, Union
+from typing import Collection, List, Tuple
 
 import numpy as np
 from rdkit import Chem
 from rdkit.Chem import AllChem
 
-PathLike = Union[str, Path]
+from ..type_hints import PathLike
 
 
 def read_geometries_from_files(file_geometries: PathLike) -> Tuple[List[Chem.rdchem.Mol], List[np.ndarray]]: