+++ swan/state/__init__.py
@@ -1,0 +1,3 @@
+from .state import StateH5
+
+__all__ = ["StateH5"]