+++ tests/test_e3nn.py
@@ -7,7 +7,7 @@ from swan.dataset import TorchGeometricGraphData
 from swan.modeller import Modeller
 from swan.modeller.models import InvariantPolynomial
 
-from .utils_test import PATH_TEST
+from .utils_test import PATH_TEST, remove_files
 
 
 def test_e3nn_equivariant():
@@ -15,7 +15,7 @@ def test_e3nn_equivariant():
     path_data = PATH_TEST / "thousand.csv"
     data = TorchGeometricGraphData(path_data, properties=["gammas"])
     net = InvariantPolynomial()
-    modeller = Modeller(net, data)
+    modeller = Modeller(net, data, replace_state=True)
     modeller.set_optimizer('Adam', lr=0.001)
     modeller.set_scheduler("StepLR", 0.1)
     modeller.data.scale_labels()
@@ -23,3 +23,4 @@ def test_e3nn_equivariant():
     expected, predicted = modeller.validate_model()
     err = torch.functional.F.mse_loss(expected, predicted)
     assert not np.isnan(err.item())
+    remove_files()